Promote /var/lib/docker to a volume
This would be dangerous in a production system, but it does allow aufs
volumes to be mounted in a docker-in-docker situation.  This is
required by the step in tasks/main.yml that performs a docker run
command.

Read more about promoting /var/lib/docker to a volume here:
https://jpetazzo.github.io/2015/09/03/do-not-use-docker-in-docker-for-ci/

For some reason bind mounting /var/run/docker.sock (mentioned in the
previous link as the preferred solution) does not seem to fix the
issue.