Bind mount /var/lib/docker
This would be dangerous in a production system, but it does allow aufs
volumes to be mounted in a docker-in-docker situation.  This is
required by the step in tasks/main.yml that performs a docker run
command.

Read more about bind mounting /var/lib/docker here:
https://jpetazzo.github.io/2015/09/03/do-not-use-docker-in-docker-for-ci/

For some reason bind mounting /var/run/docker.sock does not seem to
work.